chore: split garden into sub Components, refactor PlantTile with useRef
